requirements.txt
matplotlib
numpy
PyYAML
scipy
tensorflow